feat: 게임 로직을 훅으로 분리하여 코드 구조 개선
